fix file watches for editors renaming files
Most text editors rename the original file to FILENAME~ on first save.
After this our inotify watch would watch the backup file instead of
the real file and miss out on further updates.

To fix this, we establish a new watch on the original filename after a
rename has happened.

I hope this does not introduce a race condition where the rename has
happened but the new file has not yes been written – we'll see.